+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## Unreleased
 
-TBD
+### Features
+
+- [\#2](https://github.com/medibloc/vc-sdk/pull/2) Add `created` and `proofPurpose` field to `ProofOptions` and `Proof` structs.
+
+### Bug fixes
+
+- [\#2](https://github.com/medibloc/vc-sdk/pull/2) Fix an invalid JSON tag: `domain` -> `challenge`
+
 
 ## [v0.0.1](https://github.com/medibloc/vc-sdk/releases/tag/v0.0.1) - 2021-05-03